assets/ewm-pallet-verification-agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pyyaml==6.0.2
